langgraph-101-task-management/pyproject.toml
[project]
name = "langgraph-101-task-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.62",
    "pydantic-ai-slim[openai]>=0.0.18",
    "python-dotenv>=1.0.1",
]
